% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.rdeque.default}
\alias{as.rdeque.default}
\title{Default method for converting to an rdeque}
\usage{
\method{as.rdeque}{default}(x, ...)
}
\arguments{
\item{x}{the input to convert to an rdeque}

\item{...}{arguments to be passed to or from other methods (ignored).}
}
\value{
a filled \code{\link{rdeque}}.
}
\description{
Default method for converting to an rdeque.
}
\details{
Elements from the input (of any type) are first converted to a list with \code{\link{as.list}},
after this an rdeque of the appropriate size is created holding the elements. The element at \code{x[[1]]}
becomes the front of the rdeque. Runs in time \eqn{O(n)}, in the size of the number of elements contained in the resulting rdeque.
}
\seealso{
\code{\link{rdeque}} for info about rdeques, \code{\link{as.rdeque}} for the generic function.
}

