% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.rpqueue.default}
\alias{as.rpqueue.default}
\title{Default method for converting to an rpqueue}
\usage{
\method{as.rpqueue}{default}(x, ...)
}
\arguments{
\item{x}{the input to convert to an rpqueue.}

\item{...}{arguments to be passed to or from other methods (ignored).}
}
\value{
a filled \code{\link{rpqueue}}.
}
\description{
Default method for converting to an rpqueue.
}
\details{
Elements from the input (of any type) are first converted to a list with \code{\link{as.list}},
after this an rpqueue of the appropriate size is created holding the elements. The element at \code{x[[1]]}
becomes the front of the rpqueue. Runs in time \eqn{O(n)}.
}
\seealso{
\code{\link{rpqueue}} for info about rpqueues, \code{\link{as.rpqueue}} for the generic function.
}

