% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.rstack.default}
\alias{as.rstack.default}
\title{Default method for converting to an rstack}
\usage{
\method{as.rstack}{default}(x, ...)
}
\arguments{
\item{x}{the input to convert to an rstack.}

\item{...}{arguments to be passed to or from other methods (ignored).}
}
\value{
a filled \code{\link{rstack}}.
}
\description{
Default method for converting to an rstack.
}
\details{
Elements from the input (of any type) are first converted to a list with \code{\link{as.list}},
after this an rstack of the appropriate size is created holding the elements. The element at \code{x[[1]]}
becomes the top of the stack.
}
\seealso{
\code{\link{rstack}} for info about rstacks, \code{\link{as.rstack}} for the generic.
}

