% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.rstack.default}
\alias{as.rstack}
\alias{as.rstack.default}
\title{Default method for converting to an rstack}
\usage{
\method{as.rstack}{default}(x)

as.rstack(x)
}
\arguments{
\item{x}{Input to convert to a stack}
}
\value{
A new rstack
}
\description{
Default method for converting to an rstack

Creates a new \code{rstack} from a given input. Coerces input to a
list first using \code{as.list}, the elements of which become elements of the stack, and the
first element becoming the top of the stack.
}
\details{
O(N) in the size of the input. Note that because data frames return a list of
columns when run through \code{as.list}, running \code{as.rstack} results in a stack of
columns, rather than a stack of rows.
}
\examples{
s <- as.rstack(1:20)
print(s)

s <- as.rstack(1:200000)
print(s)

## A stack with only 5 elements, one for each column
oops <- as.rstack(iris)
print(oops)
}
\seealso{
\code{as.rstack}
}

