% Generated by roxygen2 (4.0.2): do not edit by hand
\name{empty}
\alias{empty}
\title{Check if an rstack, rdeque, or rpqueue is empty}
\usage{
empty(x, ...)
}
\arguments{
\item{x}{rstack, rdeque, or rpqueue to check.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
logical vector of length 1.
}
\description{
Check if an rstack, rdeque, or rpqueue is empty.
}
\details{
Runs in \eqn{O(1)} time for all types.
}
\examples{
s <- rstack()
print(empty(s))        ## TRUE
s <- insert_top(s, "a")
print(empty(s))        ## FALSE
}

