% Generated by roxygen2 (4.0.2): do not edit by hand
\name{empty.rpqueue}
\alias{empty.rpqueue}
\title{Check if an rpqueue is empty}
\usage{
\method{empty}{rpqueue}(x, ...)
}
\arguments{
\item{x}{the rpqueue to check.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
logical vector of length 1.
}
\description{
Check if an rpqueue is empty.
}
\details{
Runs in \eqn{O(1)} time.
}
\seealso{
\code{\link{empty}} for the generic function that can be used on rstacks, rdeques, and rpqueues.
}

