% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fixd}
\alias{fixd}
\title{Fix an rdeque}
\usage{
fixd(d, ...)
}
\arguments{
\item{d}{rdeque to fix.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
fixed, "balanced" deque.
}
\description{
Maintains the invariant that there is always something in two stacks used by
rdeques under the hood so long as there is 2 more elements in the rdeque.
}
\details{
In fact, fix will be called whenever there are fewer than 6 elements in both
the front and end of the deque. Generally this method is \eqn{O(N)}, and so a full copy is returned.
}
\references{
Okasaki, Chris. Purely Functional Data Structures. Cambridge University Press, 1999.
}

