% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fixd.rdeque}
\alias{fixd.rdeque}
\title{Fix an rdeque}
\usage{
\method{fixd}{rdeque}(d, ...)
}
\arguments{
\item{d}{The rdeque to fix.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a fixed deque.
}
\description{
A method used behind the scenes to provide \eqn{O(1)}-amortized time for most operations.
Runs in \eqn{O(n)} time worst case; restructures the rdeque so that the two internal rstacks
are roughly the same length.
}
\references{
Okasaki, Chris. Purely Functional Data Structures. Cambridge University Press, 1999.
}

