% Generated by roxygen2 (4.0.2): do not edit by hand
\name{head.rstack}
\alias{head.rstack}
\title{Return the top n elements of an rstack as an rstack.}
\usage{
\method{head}{rstack}(x, n = 6L, ...)
}
\arguments{
\item{x}{rstack to get the head of.}

\item{n}{number of elements to get.}

\item{...}{arguments to be passed to or from other methods (ignored).}
}
\value{
a smaller rstack, with only the top elements kept.
}
\description{
Returns the top n elements of a stack as a stack, or all of the elements if its length is less than n.
}
\details{
Runs in O(n) time (in the size of the number of elements requested)
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")
s <- insert_top(s, "c")

st <- head(s, n = 2)
print(st)
}

