% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_back.rdeque}
\alias{insert_back}
\alias{insert_back.rdeque}
\title{Default method for \code{rdeque} \code{insert_back}}
\usage{
\method{insert_back}{rdeque}(d, e)

insert_back(d, e)
}
\arguments{
\item{d}{The \code{rdeque} to insert onto.}

\item{e}{The element to insert.}
}
\value{
Modified version of the deque
}
\description{
Default method for \code{rdeque} \code{insert_back}

Returns a version of the deque with the new element in the back position.
}
\details{
O(1) time worst-case. Does not modify the original deque.
}
\examples{
d <- rdeque()
d <- insert_back(d, "a")
d <- insert_back(d, "b")
print(d)

d2 <- insert_back(d, "c")
print(d2)
print(d)
}
\seealso{
\code{insert_back}
}

