% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_back.rpqueue}
\alias{insert_back.rpqueue}
\title{Insert an element into the back of an rpqueue}
\usage{
\method{insert_back}{rpqueue}(x, e, ...)
}
\arguments{
\item{x}{rpqueue to insert onto.}

\item{e}{element to insert.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
modified version of the rpqueue.
}
\description{
Returns a version of the queue with the new element in the back position.
}
\details{
Runs in \eqn{O(1)} time worst-case. Does not modify the original.
}
\examples{
q <- rpqueue()
q <- insert_back(q, "a")
q <- insert_back(q, "b")
print(q)

q2 <- insert_back(q, "c")
print(q2)
print(q)
}
\references{
Okasaki, Chris. Purely Functional Data Structures. Cambridge University Press, 1999.
}

