% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_top}
\alias{insert_top}
\title{Insert an element into the top of an \code{rstack}}
\usage{
insert_top(s, e, ...)
}
\arguments{
\item{s}{rstack to insert onto.}

\item{e}{element to insert.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
modified version of the stack.
}
\description{
Returns a version of the stack with the new element in the top position.
}
\details{
O(1) time worst-case. Does not modify the original stack.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")
print(s)

s2 <- insert_top(s, "c")
print(s2)
print(s)
}

