% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_top.rstack}
\alias{insert_top.rstack}
\title{Default method for \code{insert_top} for rstack.}
\usage{
\method{insert_top}{rstack}(s, newdata)
}
\description{
Default method for \code{insert_top} for rstack.
}
\seealso{
\code{insert_top}
}

