% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_empty.rdeque}
\alias{is_empty}
\alias{is_empty.rdeque}
\alias{is_empty.rstack}
\title{Default method for checking emptyness of an rdeque}
\usage{
\method{is_empty}{rdeque}(d)

\method{is_empty}{rstack}(d)

is_empty(d)
}
\arguments{
\item{d}{The \code{rstack} or \code{rdeque} to check}
}
\value{
1-element logical vector
}
\description{
Default method for checking emptyness of an rdeque

Default method for checking emptyness of an rstack

Returns \code{TRUE} if the stack or deque has length 0, FALSE otherwise.
}
\examples{
s <- rstack()
print(is_empty(s))        ## TRUE
s <- insert_top(s, "a")
print(is_empty(s))        ## FALSE
}
\seealso{
\code{is_empty}

\code{is_empty}
}

