% Generated by roxygen2 (4.0.2): do not edit by hand
\name{length.rdeque}
\alias{length.rdeque}
\title{Default method for \code{length} on an rdeque}
\usage{
\method{length}{rdeque}(x)
}
\arguments{
\item{x}{The deque to get the length of}
}
\value{
A vector of length 1, which the number of elements of the deque
}
\description{
Returns the number of elements in an \code{rdeque}.
}
\details{
O(1) time, as this information is stored seperately and updated on insert/remove.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
print(length(d))         # 1
d <- insert_back(d, "b")
print(length(d))         # 2
}

