% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeequal.rpqueue}
\alias{makeequal.rpqueue}
\title{Maintenance function for rpqueues}
\usage{
\method{makeequal}{rpqueue}(rpqueue, ...)
}
\arguments{
\item{rpqueue}{rpqueue to makeequal.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
a "fixed" rpqueue.
}
\description{
Maintenance function for rpqueues, called automatically when needed by other functions.
}
\details{
See \emph{Simple and Efficient Purely Functional Queues and Deques},
Okasaki 1995, J. Functional Programming, 5(4) 583 to 592 for information.
}
\references{
Okasaki, Chris. Purely Functional Data Structures. Cambridge University Press, 1999.
}
\seealso{
\code{\link{rotate}} helper function that calls this one.
}

