% Generated by roxygen2 (4.0.1): do not edit by hand
\name{peek_back.rdeque}
\alias{peek_back}
\alias{peek_back.rdeque}
\title{Default method for \code{rdeque} \code{peek_back}}
\usage{
\method{peek_back}{rdeque}(d)

peek_back(d)
}
\arguments{
\item{d}{The deque to look at.}
}
\value{
The data element.
}
\description{
Default method for \code{rdeque} \code{peek_back}

Simply returns the data element sitting at the back of the deque,
leaving the deque alone.
}
\details{
O(1) worst-case time.
}
\examples{
d <- rstack()
d <- insert_front(d, "a")
d <- insert_back(d, "b")
e <- peek_back(d)
print(e)
print(d)
}
\seealso{
\code{peek_back}
}

