% Generated by roxygen2 (4.0.1): do not edit by hand
\name{peek_front.rdeque}
\alias{peek_front}
\alias{peek_front.rdeque}
\title{Default method for \code{rdeque} \code{peek_front}}
\usage{
\method{peek_front}{rdeque}(d)

peek_front(d)
}
\arguments{
\item{d}{The deque to look at.}
}
\value{
The data element.
}
\description{
Default method for \code{rdeque} \code{peek_front}

Simply returns the data element sitting at the front of the deque,
leaving the deque alone.
}
\details{
O(1) worst-case time.
}
\examples{
d <- rstack()
d <- insert_front(d, "a")
d <- insert_back(d, "b")
e <- peek_front(d)
print(e)
print(d)
}
\seealso{
\code{peek_front}
}

