% Generated by roxygen2 (4.0.1): do not edit by hand
\name{peek_top.rstack}
\alias{peek_top}
\alias{peek_top.rstack}
\title{Default method for \code{peek_top} for rstacks.}
\usage{
\method{peek_top}{rstack}(s)

peek_top(s)
}
\arguments{
\item{s}{The stack to look at}
}
\value{
The data element.
}
\description{
Default method for \code{peek_top} for rstacks.

Simply returns the data element sitting at the top of the stack,
leaving the stack alone.
}
\details{
O(1) worst-case time.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")
e <- peek_top(s)
print(e)
print(s)
}
\seealso{
\code{peek_top}
}

