% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.rdeque}
\alias{print.rdeque}
\title{Print an rdeque}
\usage{
\method{print}{rdeque}(x, ...)
}
\arguments{
\item{x}{the rdeque to print.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\description{
Prints a summary of the contents of an rdeque, including the number of elements and the first and last few.
}
\details{
Depending on the internal state of the rdeque, this method is not gauranteed to run in \eqn{O(1)} time.
}
\seealso{
\code{\link{as.list.rdeque}} for converting an rdeque into a list which can then be printed in full.
}

