% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.rstack}
\alias{print.rstack}
\title{Print an rstack}
\usage{
\method{print}{rstack}(x, ...)
}
\arguments{
\item{x}{the rstack to print.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\description{
Prints a summary of the contents of an rstack, including the number of elements and the top few.
}
\details{
Since only the top few elements are detailed, runs in \eqn{O(1)} time.
}
\seealso{
\code{\link{as.list.rstack}} for converting an rstack into a list which can then be printed in full.
}

