% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rdeque}
\alias{rdeque}
\title{Create a new rdeque}
\usage{
rdeque()
}
\description{
Creates a new, empty, \code{rdeque} ready for use.
}
\details{
An rdeque supports efficient insert into the front and back with \code{insert_front} and
\code{insert_back} (returning a version of
the deque with the new element), peek into the front and back with \code{peek_front} and \code{peek_back}
(returing the data stored at the front or back), and removal with
\code{without_front} and \code{without_back} (returning a version with the front or back
 element removed).

Other handy functions
include \code{as.list} and \code{as.data.frame} (the latter of which requires that
all elements can be appended to become rows of a data frame in a reasonable manner). Most operations
are amortized O(1) (so long as we assume the cost of removal is charged to the new variable).
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
d <- insert_back(d, "c")
d <- insert_back(d, "d")
print(d)

df <- without_back(d)
print(df)
print(d)

b <- peek_front(d)
print(b)
}

