% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rstacknode}
\alias{rstacknode}
\title{Create a new rstacknode with a given object}
\usage{
rstacknode(data)
}
\arguments{
\item{data}{Data to reference with this node}
}
\value{
Environment
}
\description{
For internal use by \code{rstack}s and \code{rdeque}s. An environment with no parent,
reference for the data and the next node.
}

