% Generated by roxygen2 (4.0.1): do not edit by hand
\name{without_top.rstack}
\alias{without_top}
\alias{without_top.rstack}
\title{Default method for \code{rstack} \code{without_top}}
\usage{
\method{without_top}{rstack}(s, n = 1)

without_top(s, n)
}
\arguments{
\item{s}{The stack to remove elements from}

\item{n}{The number of elements to remove}
}
\value{
A version of the stack with \code{n} elements removed.
}
\description{
Default method for \code{rstack} \code{without_top}

Return version of the stack without the top
}
\details{
Simply returns a version of the given stack without the top \code{n} elements
(n = 1 by default). The original stack is left alone.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")
s <- insert_top(s, "c")

s2 <- without_top(s)
print(s2)

s3 <- without_top(s, n = 2)
print(s3)

print(s)
}
\seealso{
\code{without_top}
}

