% Generated by roxygen2 (4.0.1): do not edit by hand
\name{without_top.rstack}
\alias{without_top.rstack}
\title{Default method for \code{rstack} \code{without_top}}
\usage{
\method{without_top}{rstack}(s, n = 1)
}
\description{
Default method for \code{rstack} \code{without_top}
}
\seealso{
\code{without_top}
}

