% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.rdeque}
\alias{as.data.frame.rdeque}
\title{Convert an rdeque to a data.frame, if applicable.}
\usage{
\method{as.data.frame}{rdeque}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{rdeque to convert.}

\item{row.names}{passed on to as.data.frame before final conversion.}

\item{optional}{passed onto as.data.frame before final conversion.}

\item{...}{passed onto as.data.frame before final conversion.}
}
\value{
a data.frame with the first row(s) the previous front of the deque, last row(s) previous back of the deque.
}
\description{
Converts the elements of an \code{rdeque} into rows of a dataframe, if this is reasonable.
}
\details{
This method runs in O(N) time, and will only work if all elements of the deque have the
same length() (e.g., same number of columns), and if any of the elements have names, then those
names do not conflict (e.g., same column names where used). This is accomplished by a call to
\code{do.call("rbind", as.list(d))}, where \code{as.list(d)} converts the deque \code{d} to a list
where the front element becomes the first element of the list and the back element becomes the last of the list.
}
\examples{
d <- rdeque()
d <- insert_front(d, data.frame(names = c("Bob", "Joe"), ages = c(25, 18)))
d <- insert_back(d, data.frame(names = c("Mary", "Kate", "Ashley"), ages = c(27, 26, 21)))
print(d)

ddf <- as.data.frame(d)
print(ddf)


## Building a deque in a loop, converting to a dataframe after the fact;
## sometimes adding to the front, sometimes to the back
d <- rdeque()
for(i in 1:1000) {
 if(runif(1,0,1) < 0.5) {
   d <- insert_front(d, data.frame(i = i, type = "sqrt", val = sqrt(i)))
 } else {
   d <- insert_back(d, data.frame(i = i, type = "log", val = log(i)))
 }
 if(i \%\% 100 == 0) {
   print(i/1000)
 }
}
print(head(as.data.frame(d)))
print(tail(as.data.frame(d)))
}

