% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.list.rdeque}
\alias{as.list.rdeque}
\title{Conversion of an \code{rdeque} to a \code{list}}
\usage{
\method{as.list}{rdeque}(x, ...)
}
\arguments{
\item{x}{The stack to convert}

\item{...}{Additional parameters sent to as.list after initial conversion.}
}
\value{
A list
}
\description{
Converts an \code{rdeque} to a list, where the front of the deque becomes
the first element of the list, the back the last.
}
\details{
O(N), but the generated list is pre-allocated for efficiency.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")

dlist <- as.list(d)
print(dlist)
}

