% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.list.rstack}
\alias{as.list.rstack}
\title{Conversion of an \code{rstack} to a \code{list}}
\usage{
\method{as.list}{rstack}(x, ...)
}
\arguments{
\item{x}{The stack to convert}

\item{...}{Additional arguments passed to as.list after initial conversion to list.}
}
\value{
A list
}
\description{
Converts an \code{rstack} to a list, where the top of the stack becomes
the first element of the list, the second-from-top the second, and so on.
}
\details{
O(N), but the generated list is pre-allocated for efficiency.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")

slist <- as.list(s)
print(slist)
}

