% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.rdeque}
\alias{as.rdeque}
\title{Create an \code{rdeque} from a given input}
\usage{
as.rdeque(x, ...)
}
\arguments{
\item{x}{Input to convert to a deque}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A new rdeque
}
\description{
Creates a new \code{rdeque} from a given input. Coerces input to a
list first using \code{as.list}, the elements of which become elements of the deque, and the
first element becoming the front of the deque.
}
\details{
O(N) in the size of the input. Note that because data frames return a list of
columns when run through \code{as.list}, running \code{as.rdeque} results in a deque of
columns, rather than a deque of rows.
}
\examples{
d <- as.rdeque(1:20)
print(d)

d <- as.rdeque(1:200000)
print(d)

## A deque with only 5 elements, one for each column
oops <- as.rdeque(iris)
print(oops)
}

