% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fix.rdeque}
\alias{fix}
\alias{fix.rdeque}
\title{Internal method for maintaining deque-i-ness}
\usage{
\method{fix}{rdeque}(d)

fix(d)
}
\arguments{
\item{d}{The deque to fix}
}
\value{
The fixed, "balanced" deque
}
\description{
Internal method for maintaining deque-i-ness

Maintains the invariant that there is always something in $l and $r
so long as there is 2 more elements in the queue.
}
\details{
In fact, fix will be called whenever there are fewer than 6 elements in both
the front and end of the deque. Generally this method is O(N), and so a full copy is returned.
}

