% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_back}
\alias{insert_back}
\title{Insert an element into the back of an \code{rdeque} or \code{rpqueue}}
\usage{
insert_back(x, e, ...)
}
\arguments{
\item{x}{The deque or queue to insert onto.}

\item{e}{The element to insert.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
Modified version of the deque/queue.
}
\description{
Returns a version of the deque/queue with the new element in the back position.
}
\details{
O(1) time worst-case. Does not modify the original.
}
\examples{
d <- rdeque()
d <- insert_back(d, "a")
d <- insert_back(d, "b")
print(d)

d2 <- insert_back(d, "c")
print(d2)
print(d)
}

