% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_front.rdeque}
\alias{insert_front}
\alias{insert_front.rdeque}
\title{Default method for \code{rdeque} \code{insert_front}}
\usage{
\method{insert_front}{rdeque}(d, e)

insert_front(d, e)
}
\arguments{
\item{d}{The \code{rdeque} to insert onto.}

\item{e}{The element to insert.}
}
\value{
Modified version of the deque
}
\description{
Default method for \code{rdeque} \code{insert_front}

Returns a version of the deque with the new element in the front position.
}
\details{
O(1) time worst-case. Does not modify the original deque.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
print(d)

d2 <- insert_front(d, "c")
print(d2)
print(d)
}
\seealso{
\code{insert_front}
}

