% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert_front}
\alias{insert_front}
\title{Insert an element into the front of an \code{rdeque}}
\usage{
insert_front(d, e)
}
\arguments{
\item{d}{The \code{rdeque} to insert onto.}

\item{e}{The element to insert.}
}
\value{
Modified version of the deque
}
\description{
Returns a version of the deque with the new element in the front position.
}
\details{
O(1) time worst-case. Does not modify the original deque.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
print(d)

d2 <- insert_front(d, "c")
print(d2)
print(d)
}

