% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_top.rstack}
\alias{insert_top}
\alias{insert_top.rstack}
\title{Default method for \code{insert_top} for rstack.}
\usage{
\method{insert_top}{rstack}(s, e)

insert_top(s, e)
}
\arguments{
\item{s}{The \code{rstack} to insert onto.}

\item{e}{The element to insert.}
}
\value{
Modified version of the stack.
}
\description{
Default method for \code{insert_top} for rstack.

Returns a version of the stack with the new element in the top position.
}
\details{
O(1) time worst-case. Does not modify the original stack.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")
print(s)

s2 <- insert_top(s, "c")
print(s2)
print(s)
}
\seealso{
\code{insert_top}
}

