% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_empty}
\alias{is_empty}
\title{Check if a stack or deque is empty}
\usage{
is_empty(d)
}
\arguments{
\item{d}{The \code{rstack} or \code{rdeque} to check}
}
\value{
1-element logical vector
}
\description{
Returns \code{TRUE} if the stack or deque has length 0, FALSE otherwise.
}
\examples{
s <- rstack()
print(is_empty(s))        ## TRUE
s <- insert_top(s, "a")
print(is_empty(s))        ## FALSE
}

