% Generated by roxygen2 (4.0.1): do not edit by hand
\name{length.rstack}
\alias{length.rstack}
\title{Default method for \code{length} on an rstack.}
\usage{
\method{length}{rstack}(x)
}
\arguments{
\item{x}{The stack to get the length of}
}
\value{
A vector of length 1, which the number of elements of the stack.
}
\description{
Returns the number of elements in an \code{rstack}.
}
\details{
O(1) time, as this information is stored seperately and updated on insert/remove.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
print(length(s))         # 1
s <- insert_top(s, "b")
print(length(s))         # 2
}

