% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peek_back}
\alias{peek_back}
\title{Return the data element at the back of an rdeque.}
\usage{
peek_back(d, ...)
}
\arguments{
\item{d}{rdeque to look at.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
data element at the back of the rdeque.
}
\description{
Simply returns the data element sitting at the back of the deque,
leaving the deque alone.
}
\details{
O(1) worst-case time.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_back(d, "b")
e <- peek_back(d)
print(e)
print(d)
}

