% Generated by roxygen2 (4.0.1): do not edit by hand
\name{peek_back}
\alias{peek_back}
\title{Return the data element at the back of the deque.}
\usage{
peek_back(d)
}
\arguments{
\item{d}{The deque to look at.}
}
\value{
The data element.
}
\description{
Simply returns the data element sitting at the back of the deque,
leaving the deque alone.
}
\details{
O(1) worst-case time.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_back(d, "b")
e <- peek_back(d)
print(e)
print(d)
}

