% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peek_front.R
\name{peek_front}
\alias{peek_front}
\title{Return the data element at the front of an rdeque}
\usage{
peek_front(x, ...)
}
\arguments{
\item{x}{rdeque to look at.}

\item{...}{additional arguments to be passed to or from methods (ignored).}
}
\value{
data element at the front of the rdeque.
}
\description{
Simply returns the data element sitting at the front of the deque,
leaving the deque alone.
}
\details{
Runs in \eqn{O(1)} worst-case time.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_back(d, "b")
e <- peek_front(d)
print(e)
print(d)
}

