% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peek_front}
\alias{peek_front}
\title{Return the data element at the front of the deque.}
\usage{
peek_front(x, ...)
}
\arguments{
\item{x}{The deque to look at.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
The data element.
}
\description{
Simply returns the data element sitting at the front of the deque,
leaving the deque alone.
}
\details{
O(1) worst-case time.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_back(d, "b")
e <- peek_front(d)
print(e)
print(d)
}

