% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.rdeque}
\alias{print.rdeque}
\title{Default print method for an rdeque}
\usage{
\method{print}{rdeque}(x, ...)
}
\arguments{
\item{x}{The deque to print}

\item{...}{Arguments passed to or from other functions (unused)}
}
\description{
Prints the front and back of an rdeque, up to the first 6 elements each.
}

