% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.rstack}
\alias{print.rstack}
\title{Default print method for an rstack}
\usage{
\method{print}{rstack}(x, ...)
}
\arguments{
\item{x}{The stack to print}

\item{...}{Arguments to be passed to or from other methods (ignored)}
}
\description{
Prints the top of an rstack, up to the first 6 elements.
}

