% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rstacknode}
\alias{rstacknode}
\title{Create a new rstacknode from a given object.}
\usage{
rstacknode(data)
}
\arguments{
\item{data}{data to reference with this node.}
}
\value{
an environment.
}
\description{
For use by rstacks and rdeques. Simply an environment with no parent,
reference for the data and the next node.
}
\details{
A class providing an O(1) worst-case time for stacks. Exported functions:

rstack()
as.rstack(input)
empty(s)
print(s)
as.data.frame(s)
peek_top(s)
length(s)
insert_top(s, element)
without_top(s)
as.list(s)
rev(s)
head(s, n = 6)
}

