% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rstacknode}
\alias{rstacknode}
\title{Internal structure used by rstacks, rdeques, and rpqueues.}
\usage{
rstacknode(data)
}
\arguments{
\item{data}{data to reference with this node.}
}
\value{
an environment.
}
\description{
For use by rstacks and rdeques. Simply an environment with no parent,
reference for the data and the next node.
}

