% Generated by roxygen2 (4.0.2): do not edit by hand
\name{without_back}
\alias{without_back}
\title{Return a version of an rdeque without the back element.}
\usage{
without_back(d, ...)
}
\arguments{
\item{d}{rdeque to remove elements from.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a version of the rdeque with \code{n} elements removed.
}
\description{
Return a version of an rdeque without the back element.
}
\details{
Simply returns a version of the given deque without the back \code{n} elements
(n = 1 by default). The original deque is left alone.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
d <- insert_back(d, "c")

d2 <- without_back(d)
print(d2)

d3 <- without_back(d2)
print(d3)

print(d)
}

