% Generated by roxygen2 (4.0.1): do not edit by hand
\name{without_front}
\alias{without_front}
\title{Return version of a deque/queue without the front}
\usage{
without_front(x)
}
\arguments{
\item{x}{The deque/queue to remove elements from}
}
\value{
A version of the deque/queue with the front element removed.
}
\description{
Return version of a deque/queue without the front
}
\details{
Simply returns a version of the given structure without the front element.
The original is left alone.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
d <- insert_back(d, "c")

d2 <- without_front(d)
print(d2)

d3 <- without_front(d2)
print(d3)

print(d)
}

