% Generated by roxygen2 (4.0.1): do not edit by hand
\name{without_front.rdeque}
\alias{without_front}
\alias{without_front.rdeque}
\title{Default method for \code{rdeque} \code{without_front}}
\usage{
\method{without_front}{rdeque}(d, n = 1)

without_front(d)
}
\arguments{
\item{d}{The deque to remove elements from}

\item{n}{The number of elements to remove}
}
\value{
A version of the deque with \code{n} elements removed.
}
\description{
Default method for \code{rdeque} \code{without_front}

Return version of the deque without the front
}
\details{
Simply returns a version of the given deque without the front \code{n} elements
(n = 1 by default). The original deque is left alone.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
d <- insert_back(d, "c")

d2 <- without_front(d)
print(d2)

d3 <- without_front(d, n = 2)
print(d3)

print(d)
}
\seealso{
\code{without_front}
}

