% Generated by roxygen2 (4.0.2): do not edit by hand
\name{without_front}
\alias{without_front}
\title{Return version of a deque/queue without the front}
\usage{
without_front(x, ...)
}
\arguments{
\item{x}{rdeque or rpqueue to remove elements from.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a version of the rdeque or rpqueue with the front element removed.
}
\description{
Return version of a deque/queue without the front
}
\details{
Simply returns a version of the given structure without the front element.
The original is left alone.
}
\examples{
d <- rdeque()
d <- insert_front(d, "a")
d <- insert_front(d, "b")
d <- insert_back(d, "c")

d2 <- without_front(d)
print(d2)

d3 <- without_front(d2)
print(d3)

print(d)
}

