% Generated by roxygen2 (4.0.2): do not edit by hand
\name{without_top}
\alias{without_top}
\title{Return version of the stack without the top}
\usage{
without_top(s, n, ...)
}
\arguments{
\item{s}{rstack to remove elements from.}

\item{n}{number of elements to remove.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
version of the stack with \code{n} elements removed.
}
\description{
Return version of the stack without the top
}
\details{
Simply returns a version of the given stack without the top \code{n} elements
(n = 1 by default). The original stack is left alone.
}
\examples{
s <- rstack()
s <- insert_top(s, "a")
s <- insert_top(s, "b")
s <- insert_top(s, "c")

s2 <- without_top(s)
print(s2)

s3 <- without_top(s, n = 2)
print(s3)

print(s)
}

